requirement.txt
openai
groq
google-generativeai
mistralai

python-dotenv
requests

fastapi
uvicorn

pydantic

langchain
langchain-community
langchain-openai


tiktoken

numpy
pandas